requirements.txt
# requirements.txt
praw
pandas
spacy
geopy
folium
vaderSentiment
plotly
scikit-learn
transformers
python-dotenv
flask
textblob
torch
seaborn
matplotlib